% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{glass}
\alias{glass}
\title{Multilingual sample text}
\format{
A named character vector, whose elements are translations of the
  sentence, and are named for the appropriate language in each case.
}
\source{
The translations were gathered by Frank da Cruz and written by a
  large group of contributors. Notes, commentary and a full list of credits
  are online at \url{https://kermitproject.org/utf8.html}.
}
\usage{
glass
}
\description{
This dataset contains translations into many languages of the esoteric
sentence "I can eat glass and it doesn't hurt me", UTF-8 encoded. Since
this dataset uses characters from a range of scripts, it provides a useful
test set for text handling and character encodings.
}
\keyword{datasets}
