% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ore.R
\name{ore.escape}
\alias{ore.escape}
\alias{ore_escape}
\title{Escape regular expression special characters}
\usage{
ore.escape(text)
}
\arguments{
\item{text}{A character vector.}
}
\value{
A modified version of the argument, with special characters escaped
  by prefixing them with a backslash.
}
\description{
Escape characters that would usually be interpreted specially in a regular
expression, returning a modified version of the argument. This can be
useful when incorporating a general-purpose string into a larger regex.
}
\seealso{
\code{\link{ore}}
}

