% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{ore.file}
\alias{ore.file}
\alias{orefile}
\alias{ore_file}
\title{Use a file as a text source}
\usage{
ore.file(path, encoding = getOption("ore.encoding"), binary = FALSE)
}
\arguments{
\item{path}{A character string giving the file path.}

\item{encoding}{A character string giving the encoding of the file. This
should match the encoding of the regular expression used in a call to
\code{\link{ore.search}}.}

\item{binary}{A logical value: if \code{TRUE}, the file will be search
bytewise, and \code{encoding} will be fixed to be \code{"ASCII"}.}
}
\description{
Identify a file path to be used as a text source for a subsequent call to
\code{\link{ore.search}}.
}
\seealso{
\code{\link{ore.search}} for actually searching through the file.
}
