% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ore.lastmatch}
\alias{ore.lastmatch}
\title{Retrieve the last match}
\usage{
ore.lastmatch()
}
\value{
An \code{"orematch"} object or list. See \code{\link{ore.search}}
  for details.
}
\description{
This function can be used to obtain the \code{"orematch"} object, or list,
corresponding to the last call to \code{\link{ore.search}}. This can be
useful after performing a search implicitly, for example with \code{\%~\%}.
}

