% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{ore_file}
\alias{ore_file}
\alias{orefile}
\alias{ore.file}
\title{Use a file as a text source}
\usage{
ore_file(path, encoding = getOption("ore.encoding"), binary = FALSE)
}
\arguments{
\item{path}{A character string giving the file path.}

\item{encoding}{A character string giving the encoding of the file. This
should match the encoding of the regular expression used in a call to
\code{\link{ore_search}}.}

\item{binary}{A logical value: if \code{TRUE}, the file will be search
bytewise, and \code{encoding} will be fixed to be \code{"ASCII"}.}
}
\value{
A string of class \code{"orefile"}, with the \code{encoding} and
  \code{binary} arguments stored as attributes.
}
\description{
Identify a file path to be used as a text source for a subsequent call to
\code{\link{ore_search}}, or to \code{\link{ore_ismatch}},
\code{\link{ore_split}}, \code{\link{ore_subst}} or \code{\link{ore_switch}}.
}
\seealso{
\code{\link{ore_search}} for actually searching through the file.
}
