% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{ore_lastmatch}
\alias{ore_lastmatch}
\alias{ore.lastmatch}
\title{Retrieve the last match}
\usage{
ore_lastmatch(simplify = TRUE)
}
\arguments{
\item{simplify}{If \code{TRUE} and the last match was against a single
string, then the \code{"orematch"} object will be returned, instead of a
list with one element.}
}
\value{
An \code{"orematch"} object or list. See \code{\link{ore_search}}
  for details.
}
\description{
This function can be used to obtain the \code{"orematch"} object, or list,
corresponding to the last call to \code{\link{ore_search}}. This can be
useful after performing a search implicitly, for example with \code{\%~\%}.
}
