% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matches}
\alias{groups}
\alias{groups.default}
\alias{groups.list}
\alias{groups.orematch}
\alias{matches}
\alias{matches.default}
\alias{matches.list}
\alias{matches.orematch}
\title{Extract matching substrings}
\usage{
matches(object, ...)

\method{matches}{list}(object, ...)

\method{matches}{orematch}(object, ...)

\method{matches}{default}(object, ...)

groups(object, ...)

\method{groups}{list}(object, ...)

\method{groups}{orematch}(object, ...)

\method{groups}{default}(object, ...)
}
\arguments{
\item{object}{An R object. Methods are provided for generic lists and
\code{"orematch"} objects.}

\item{...}{Further arguments to methods. Unused here.}
}
\value{
A vector, matrix, array, or list of the same, containing full
  matches or subgroups.
}
\description{
These functions extract entire matches, or just subgroup matches, from
objects of class \code{"orematch"}. They can also be applied to lists of
these objects, as returned by \code{\link{ore.search}} when more than one
string is searched. For other objects they return \code{NA}.
}
\seealso{
\code{\link{ore.search}}
}

